# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OC-Facturer API
  description: API de gestion de la facturation des bons de commande Sewan (OC-DB + Sophia GraphQL)
  version: 0.1.0
  contact:
    name: Dominique Pineau
    email: dominique.pineau@sewan.fr
servers:
  - url: http://localhost:3000
    description: Developpement local
  - url: https://oc-facturer.sewan.fr
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token (1h). Obtenu via POST /api/auth/login.
...
security:
  - BearerAuth: []
paths:
...
  /api/sophia/sites/select:
    post:
      tags: [Sophia]
      summary: Selectionner un site pour une ressource
      description: Verifie la hierarchie du site dans le contrat, met a jour la ressource avec le site Sophia.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [siteId, resourceId, contractId]
              properties:
                siteId:
                  type: string
                  format: uuid
                resourceId:
                  type: string
                  format: uuid
                contractId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Site associe a la ressource
  /api/sophia/sites/verify:
    post:
      tags: [Sophia]
      summary: Verifier un site dans un contrat
      parameters:
        - name: siteId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: contractId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Verification effectuee
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  verified:
                    type: boolean
...